+++ hyperspy/_signals/signal2d.py
@@ -25,9 +25,8 @@ import numpy as np
 import numpy.ma as ma
 import scipy
 
+from hyperspy import _lazy_signals, signal_tools, signals
 from hyperspy._signals.common_signal2d import CommonSignal2D
-from hyperspy._signals.lazy import LazySignal
-from hyperspy._signals.signal1d import Signal1D
 from hyperspy.defaults_parser import preferences
 from hyperspy.docstrings.plot import (
     BASE_PLOT_DOCSTRING,
@@ -41,10 +40,9 @@ from hyperspy.docstrings.signal import (
     SHOW_PROGRESSBAR_ARG,
 )
 from hyperspy.external.progressbar import progressbar
+from hyperspy.misc import utils
 from hyperspy.misc.math_tools import antisymmetrize, optimal_fft_size, symmetrize
-from hyperspy.misc.utils import is_dask_array
 from hyperspy.signal import BaseSignal
-from hyperspy.signal_tools import PeaksFinder2D, Signal2DCalibration
 from hyperspy.ui_registry import DISPLAY_DT, TOOLKIT_DT
 
 _logger = logging.getLogger(__name__)
@@ -207,7 +205,7 @@ def estimate_image_shift(
     """
     import matplotlib.pyplot as plt
 
-    if is_dask_array(ref) or is_dask_array(image):
+    if utils.is_dask_array(ref) or utils.is_dask_array(image):
         import dask.array as da
 
         ref, image = da.compute(ref, image)
@@ -726,7 +724,7 @@ class Signal2D(BaseSignal, CommonSignal2D):
             )
             return None
         if isinstance(shifts, np.ndarray):
-            signal_shifts = Signal1D(-shifts)
+            signal_shifts = signals.Signal1D(-shifts)
         else:
             signal_shifts = shifts
         if expand:
@@ -863,7 +861,7 @@ class Signal2D(BaseSignal, CommonSignal2D):
         """
         self._check_signal_dimension_equals_two()
         if interactive:
-            calibration = Signal2DCalibration(self)
+            calibration = signal_tools.Signal2DCalibration(self)
             calibration.gui(display=display, toolkit=toolkit)
         else:
             if None in (x0, y0, x1, y1, new_length):
@@ -1096,7 +1094,9 @@ class Signal2D(BaseSignal, CommonSignal2D):
             peaks = BaseSignal(
                 np.empty(self.axes_manager.navigation_shape), axes=axes_dict
             )
-            pf2D = PeaksFinder2D(self, method=method, peaks=peaks, **kwargs)
+            pf2D = signal_tools.PeaksFinder2D(
+                self, method=method, peaks=peaks, **kwargs
+            )
             pf2D.gui(display=display, toolkit=toolkit)
         elif current_index:
             peaks = method_func(self._get_current_data(), **kwargs)
@@ -1121,7 +1121,7 @@ class Signal2D(BaseSignal, CommonSignal2D):
     )
 
 
-class LazySignal2D(LazySignal, Signal2D):
+class LazySignal2D(_lazy_signals.LazySignal, Signal2D):
     """Lazy general 2D signal class."""
 
     __doc__ += LAZYSIGNAL_DOC.replace("__BASECLASS__", "Signal2D")
